Crate google_analyticsreporting4

This documentation was generated from AnalyticsReporting crate version 1.0.9+20190625, where 20190625 is the exact revision of the analyticsreporting:v4 schema built by the mako code generator v1.0.9.

Everything else about the AnalyticsReporting v4 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

reports

batch get

user activity

search

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.user_activity().search(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-analyticsreporting4 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_analyticsreporting4 as analyticsreporting4;
use analyticsreporting4::SearchUserActivityRequest;
use analyticsreporting4::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use analyticsreporting4::AnalyticsReporting;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AnalyticsReporting::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = SearchUserActivityRequest::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.user_activity().search(req)
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

Activity
An Activity represents data for an activity of a user. Note that an Activity is different from a hit. A hit might result in multiple Activity's. For example, if a hit includes a transaction and a goal completion, there will be two Activity protos for this hit, one for ECOMMERCE and one for GOAL. Conversely, multiple hits can also construct one Activity. In classic e-commerce, data for one transaction might be sent through multiple hits. These hits will be merged into one ECOMMERCE Activity.

AnalyticsReporting
Central instance to access all AnalyticsReporting related resource activities

Chunk

Cohort
Defines a cohort. A cohort is a group of users who share a common characteristic. For example, all users with the same acquisition date belong to the same cohort.

CohortGroup
Defines a cohort group. For example:

ColumnHeader
Column headers.

ContentRange
Implements the Content-Range header, for serialization only

CustomDimension
Custom dimension.

DateRange
A contiguous set of days: startDate, startDate + 1 day, ..., endDate. The start and end dates are specified in ISO8601 date format YYYY-MM-DD.

DateRangeValues
Used to return a list of metrics for a single DateRange / dimension combination

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

Dimension
Dimensions are attributes of your data. For example, the dimension ga:city indicates the city, for example, "Paris" or "New York", from which a session originates.

DimensionFilter
Dimension filter specifies the filtering options on a dimension.

DimensionFilterClause
A group of dimension filters. Set the operator value to specify how the filters are logically combined.

DummyNetworkStream

DynamicSegment
Dynamic segment definition for defining the segment within the request. A segment can select users, sessions or both.

EcommerceData
E-commerce details associated with the user activity.

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

EventData
Represents all the details pertaining to an event.

GetReportsRequest
The batch request containing multiple report request.

GetReportsResponse
The main response class which holds the reports from the Reporting API batchGet call.

GoalData
Represents all the details pertaining to a goal.

GoalSetData
Represents a set of goals that were reached in an activity.

JsonServerError
A utility type which can decode a server response that indicates error

MethodInfo
Contains information about an API request.

Metric
Metrics are the quantitative measurements. For example, the metric ga:users indicates the total number of users for the requested time period.

MetricFilter
MetricFilter specifies the filter on a metric.

MetricFilterClause
Represents a group of metric filters. Set the operator value to specify how the filters are logically combined.

MetricHeader
The headers for the metrics.

MetricHeaderEntry
Header for the metrics.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

OrFiltersForSegment
A list of segment filters in the OR group are combined with the logical OR operator.

OrderBy
Specifies the sorting options.

PageviewData
Represents details collected when the visitor views a page.

Pivot
The Pivot describes the pivot section in the request. The Pivot helps rearrange the information in the table for certain reports by pivoting your data on a second dimension.

PivotHeader
The headers for each of the pivot sections defined in the request.

PivotHeaderEntry
The headers for the each of the metric column corresponding to the metrics requested in the pivots section of the response.

PivotValueRegion
The metric values in the pivot region.

ProductData
Details of the products in an e-commerce transaction.

RangeResponseHeader

Report
The data response corresponding to the request.

ReportBatchGetCall
Returns the Analytics data.

ReportData
The data part of the report.

ReportMethods
A builder providing access to all methods supported on report resources. It is not used directly, but through the AnalyticsReporting hub.

ReportRequest
The main request class which specifies the Reporting API request.

ReportRow
A row in the report.

ResourceQuotasRemaining
The resource quota tokens remaining for the property after the request is completed.

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

ScreenviewData
There is no detailed description.

SearchUserActivityRequest
The request to fetch User Report from Reporting API userActivity:get call.

SearchUserActivityResponse
The response from userActivity:get call.

Segment
The segment definition, if the report needs to be segmented. A Segment is a subset of the Analytics data. For example, of the entire set of users, one Segment might be users from a particular country or city.

SegmentDefinition
SegmentDefinition defines the segment to be a set of SegmentFilters which are combined together with a logical AND operation.

SegmentDimensionFilter
Dimension filter specifies the filtering options on a dimension.

SegmentFilter
SegmentFilter defines the segment to be either a simple or a sequence segment. A simple segment condition contains dimension and metric conditions to select the sessions or users. A sequence segment condition can be used to select users or sessions based on sequential conditions.

SegmentFilterClause
Filter Clause to be used in a segment definition, can be wither a metric or a dimension filter.

SegmentMetricFilter
Metric filter to be used in a segment filter clause.

SegmentSequenceStep
A segment sequence definition.

SequenceSegment
Sequence conditions consist of one or more steps, where each step is defined by one or more dimension/metric conditions. Multiple steps can be combined with special sequence operators.

ServerError

ServerMessage

SimpleSegment
A Simple segment conditions consist of one or more dimension/metric conditions that can be combined.

TransactionData
Represents details collected when the visitor performs a transaction on the page.

User
Contains information to identify a particular user uniquely.

UserActivityMethods
A builder providing access to all methods supported on userActivity resources. It is not used directly, but through the AnalyticsReporting hub.

UserActivitySearchCall
Returns User Activity data.

UserActivitySession
This represents a user session performed on a specific device at a certain time over a period of time.

XUploadContentType
The X-Upload-Content-Type header.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.