Module google_analyticsreporting4::api

Structs

Activity
An Activity represents data for an activity of a user. Note that an Activity is different from a hit. A hit might result in multiple Activity's. For example, if a hit includes a transaction and a goal completion, there will be two Activity protos for this hit, one for ECOMMERCE and one for GOAL. Conversely, multiple hits can also construct one Activity. In classic e-commerce, data for one transaction might be sent through multiple hits. These hits will be merged into one ECOMMERCE Activity.

AnalyticsReporting
Central instance to access all AnalyticsReporting related resource activities

Cohort
Defines a cohort. A cohort is a group of users who share a common characteristic. For example, all users with the same acquisition date belong to the same cohort.

CohortGroup
Defines a cohort group. For example: "cohortGroup": { "cohorts": [{ "name": "cohort 1", "type": "FIRST_VISIT_DATE", "dateRange": { "startDate": "2015-08-01", "endDate": "2015-08-01" } },{ "name": "cohort 2" "type": "FIRST_VISIT_DATE" "dateRange": { "startDate": "2015-07-01", "endDate": "2015-07-01" } }] }

ColumnHeader
Column headers.

CustomDimension
Custom dimension.

DateRange
A contiguous set of days: startDate, startDate + 1 day, ..., endDate. The start and end dates are specified in ISO8601 date format YYYY-MM-DD.

DateRangeValues
Used to return a list of metrics for a single DateRange / dimension combination

Dimension
Dimensions are attributes of your data. For example, the dimension ga:city indicates the city, for example, "Paris" or "New York", from which a session originates.

DimensionFilter
Dimension filter specifies the filtering options on a dimension.

DimensionFilterClause
A group of dimension filters. Set the operator value to specify how the filters are logically combined.

DynamicSegment
Dynamic segment definition for defining the segment within the request. A segment can select users, sessions or both.

EcommerceData
E-commerce details associated with the user activity.

EventData
Represents all the details pertaining to an event.

GetReportsRequest
The batch request containing multiple report request.

GetReportsResponse
The main response class which holds the reports from the Reporting API batchGet call.

GoalData
Represents all the details pertaining to a goal.

GoalSetData
Represents a set of goals that were reached in an activity.

Metric
Metrics are the quantitative measurements. For example, the metric ga:users indicates the total number of users for the requested time period.

MetricFilter
MetricFilter specifies the filter on a metric.

MetricFilterClause
Represents a group of metric filters. Set the operator value to specify how the filters are logically combined.

MetricHeader
The headers for the metrics.

MetricHeaderEntry
Header for the metrics.

OrFiltersForSegment
A list of segment filters in the OR group are combined with the logical OR operator.

OrderBy
Specifies the sorting options.

PageviewData
Represents details collected when the visitor views a page.

Pivot
The Pivot describes the pivot section in the request. The Pivot helps rearrange the information in the table for certain reports by pivoting your data on a second dimension.

PivotHeader
The headers for each of the pivot sections defined in the request.

PivotHeaderEntry
The headers for the each of the metric column corresponding to the metrics requested in the pivots section of the response.

PivotValueRegion
The metric values in the pivot region.

ProductData
Details of the products in an e-commerce transaction.

Report
The data response corresponding to the request.

ReportBatchGetCall
Returns the Analytics data.

ReportData
The data part of the report.

ReportMethods
A builder providing access to all methods supported on report resources. It is not used directly, but through the AnalyticsReporting hub.

ReportRequest
The main request class which specifies the Reporting API request.

ReportRow
A row in the report.

ResourceQuotasRemaining
The resource quota tokens remaining for the property after the request is completed.

ScreenviewData
There is no detailed description.

SearchUserActivityRequest
The request to fetch User Report from Reporting API userActivity:get call.

SearchUserActivityResponse
The response from userActivity:get call.

Segment
The segment definition, if the report needs to be segmented. A Segment is a subset of the Analytics data. For example, of the entire set of users, one Segment might be users from a particular country or city.

SegmentDefinition
SegmentDefinition defines the segment to be a set of SegmentFilters which are combined together with a logical AND operation.

SegmentDimensionFilter
Dimension filter specifies the filtering options on a dimension.

SegmentFilter
SegmentFilter defines the segment to be either a simple or a sequence segment. A simple segment condition contains dimension and metric conditions to select the sessions or users. A sequence segment condition can be used to select users or sessions based on sequential conditions.

SegmentFilterClause
Filter Clause to be used in a segment definition, can be wither a metric or a dimension filter.

SegmentMetricFilter
Metric filter to be used in a segment filter clause.

SegmentSequenceStep
A segment sequence definition.

SequenceSegment
Sequence conditions consist of one or more steps, where each step is defined by one or more dimension/metric conditions. Multiple steps can be combined with special sequence operators.

SimpleSegment
A Simple segment conditions consist of one or more dimension/metric conditions that can be combined.

TransactionData
Represents details collected when the visitor performs a transaction on the page.

User
Contains information to identify a particular user uniquely.

UserActivityMethods
A builder providing access to all methods supported on userActivity resources. It is not used directly, but through the AnalyticsReporting hub.

UserActivitySearchCall
Returns User Activity data.

UserActivitySession
This represents a user session performed on a specific device at a certain time over a period of time.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.